requirements.txt
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
idna==2.8
Mastodon.py==1.4.5
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
